Insert map, fixed
